requirements.txt
Package                Version
---------------------- ------------------
absl-py                0.11.0
aio-pika               6.4.1
aiormq                 3.3.1
alembic                1.4.3
aniso8601              8.0.0
appdirs                1.4.4
argon2-cffi            20.1.0
astor                  0.8.1
astunparse             1.6.3
async-generator        1.10
attrs                  20.3.0
backcall               0.2.0
bert-dp                1.0
bleach                 3.2.1
blis                   0.4.1
boto3                  1.16.23
botocore               1.19.23
cachetools             4.1.1
catalogue              1.0.0
certifi                2020.11.8
cffi                   1.14.3
chardet                3.0.4
click                  7.1.2
cloudpickle            1.6.0
colorama               0.4.4
coverage               5.3
cryptography           3.2.1
cymem                  1.31.2
Cython                 0.29.14
cytoolz                0.9.0.1
databricks-cli         0.14.1
dataclasses            0.6
datasets               1.1.2
DAWG-Python            0.7.2
decorator              4.4.2
deeppavlov             0.13.0
defusedxml             0.6.0
dill                   0.2.9
distlib                0.3.1
docker                 4.3.1
docopt                 0.6.2
dotmap                 1.3.0
elastic-apm            5.10.0
elasticsearch          7.10.0
en-core-web-md         2.0.0
en-core-web-sm         2.3.1
en-coref-md            3.0.0
entrypoints            0.3
farm                   0.5.0
farm-haystack          0.5.0
fastapi                0.47.1
fastpunct              1.0.2
filelock               3.0.12
Flask                  1.1.2
Flask-Cors             3.0.9
Flask-MySQLdb          0.2.0
flask-restplus         0.13.0
future                 0.18.2
gast                   0.3.3
gdown                  3.12.2
gitdb                  4.0.5
GitPython              3.1.11
google-auth            1.23.0
google-auth-oauthlib   0.4.2
google-pasta           0.2.0
grpcio                 1.33.2
gunicorn               20.0.4
h11                    0.9.0
h5py                   2.10.0
httptools              0.1.1
idna                   2.8
importlib-metadata     2.0.0
ipykernel              5.3.4
ipython                7.19.0
ipython-genutils       0.2.0
ipywidgets             7.5.1
itsdangerous           1.1.0
jedi                   0.17.2
Jinja2                 2.11.2
jmespath               0.10.0
joblib                 0.17.0
jsonschema             3.2.0
jupyter                1.0.0
jupyter-client         6.1.7
jupyter-console        6.2.0
jupyter-core           4.7.0
jupyterlab-pygments    0.1.2
Keras                  2.4.3
Keras-Applications     1.0.8
Keras-Preprocessing    1.1.2
langdetect             1.0.8
lxml                   4.6.1
Mako                   1.1.3
Markdown               3.3.3
MarkupSafe             1.1.1
mistune                0.8.4
mlflow                 1.0.0
more-itertools         8.6.0
msgpack                0.6.2
msgpack-numpy          0.4.7.1
multidict              5.0.2
multiprocess           0.70.11.1
murmurhash             0.28.0
mysql-connector-python 8.0.22
mysqlclient            2.0.1
nbclient               0.5.1
nbconvert              6.0.7
nbformat               5.0.8
nest-asyncio           1.4.3
networkx               2.5
nltk                   3.4.5
notebook               6.1.5
numpy                  1.18.0
oauthlib               3.1.0
opt-einsum             3.3.0
overrides              2.7.0
packaging              20.4
pamqp                  2.3.0
pandas                 0.25.3
pandocfilters          1.4.3
parso                  0.7.1
pickleshare            0.7.5
Pillow                 8.0.1
pip                    20.1.1
plac                   0.9.6
pluggy                 0.13.1
preshed                1.0.1
progressbar2           3.53.1
prometheus-client      0.7.1
prompt-toolkit         3.0.8
protobuf               3.14.0
psutil                 5.7.3
punctuator             0.9.6
py                     1.9.0
pyarrow                2.0.0
pyasn1                 0.4.8
pyasn1-modules         0.2.8
pycparser              2.20
pydantic               1.3
pydload                1.0.8
Pygments               2.7.2
pymorphy2              0.8
pymorphy2-dicts        2.4.393442.3710985
pymorphy2-dicts-ru     2.4.417127.4579844
pyOpenSSL              19.1.0
pyparsing              2.4.7
PyPDF2                 1.26.0
pyrsistent             0.17.3
PySocks                1.7.1
pyTelegramBotAPI       3.6.7
python-dateutil        2.8.1
python-docx            0.8.10
python-editor          1.0.4
python-multipart       0.0.5
python-utils           2.4.0
pytz                   2019.1
pywin32                227
pywinpty               0.5.7
PyYAML                 5.3.1
pyzmq                  20.0.0
qtconsole              4.7.7
QtPy                   1.9.0
querystring-parser     1.2.4
regex                  2017.4.5
requests               2.22.0
requests-oauthlib      1.3.0
rsa                    4.6
ruamel.yaml            0.15.100
rusenttokenize         0.0.5
s3transfer             0.3.3
sacremoses             0.0.35
scikit-learn           0.21.2
scipy                  1.4.1
Send2Trash             1.5.0
sentencepiece          0.1.91
seqeval                0.0.12
setuptools             47.1.0
simplejson             3.17.2
six                    1.15.0
sklearn                0.0
smmap                  3.0.4
spacy                  2.0.12
SQLAlchemy             1.3.20
SQLAlchemy-Utils       0.36.8
sqlparse               0.4.1
srsly                  1.0.4
starlette              0.12.9
tabulate               0.8.7
tensorboard            2.4.0
tensorboard-plugin-wit 1.7.0
tensorflow             2.3.1
tensorflow-estimator   2.3.0
termcolor              1.1.0
terminado              0.9.1
testpath               0.4.4
Theano                 1.0.5
thinc                  6.10.3
tika                   1.24
tokenizers             0.9.3
toml                   0.10.2
toolz                  0.11.1
torch                  1.6.0+cu101
torchaudio             0.7.0
torchvision            0.6.1+cu101
tornado                6.1
tox                    3.20.1
tqdm                   4.41.1
traitlets              5.0.5
transformers           3.5.1
typing-extensions      3.7.4.3
ujson                  4.0.1
urllib3                1.25.4
uvicorn                0.11.7
virtualenv             20.2.0
wasabi                 0.8.0
wcwidth                0.2.5
webencodings           0.5.1
websocket-client       0.57.0
websockets             8.1
Werkzeug               0.16.1
wheel                  0.35.1
widgetsnbextension     3.5.1
wrapt                  1.12.1
xxhash                 2.0.0
yarl                   1.6.3
zipp                   3.4.0